// isem-finanzas-web/src/app/@finanzas/@auditorias/models/adjunto.model.ts
export class AdjuntoModel {
  id_adjunto?: number;
  nombre?: string;
  tamanio?: string;
  fecha_carga?: string;
  activo?: boolean;
  id_tipo_adjunto?: number;
  tipo_nombre?: string;
  id_oficio?: number;
  id_requerimiento_auditoria?: number;
  id_suficiencia?: number;
  base64?: string;
  tipo?: string;
}